+++ lib/src/models/msg_model.dart
@@ -1,4 +1,5 @@
 import 'package:swift_agents/src/models/conversations_response.dart';
+import 'package:swift_agents/src/models/upload_attachments_response.dart';
 
 import '../screens/widgets/chat_bubble.dart';
 
@@ -6,7 +7,8 @@ class MsgModel {
   final String text;
   final BubbleRole role;
   final ConversationSession? session;
+  final List<AttachmentModel>? attachments;
 
-  MsgModel(this.text, this.role, {this.session});
+  MsgModel(this.text, this.role, this.attachments, {this.session});
 }