+++ lib/src/utils/file_util.dart
@@ -9,6 +9,26 @@ import 'package:uuid/uuid.dart';
 import 'package:crypto/crypto.dart';
 
 class FileUtils {
+    static String getFileNameFromSignature(String signature) {
+    const prefix = 'sw_';
+
+    if (!signature.startsWith(prefix)) {
+      throw ArgumentError('Invalid signature');
+    }
+
+    // Remove "sw_"
+    final withoutPrefix = signature.substring(prefix.length);
+
+    // Find the first "_" after the hash
+    final firstUnderscore = withoutPrefix.indexOf('_');
+
+    if (firstUnderscore == -1) {
+      throw ArgumentError('Invalid signature');
+    }
+
+    return withoutPrefix.substring(firstUnderscore + 1);
+  }
+
   String getFileSignature(String name, Uint8List? bytes) {
     String hash = bytes != null
         ? sha256.convert(bytes).toString()
@@ -39,7 +59,7 @@ class FileUtils {
     }
   }
 
-
+  // PICKERS
   Future<List<UploadFile>?> imagesPicker(BuildContext context) async {
     final permissionsProvider = Provider.of<PermissionsProvider>(
       context,
@@ -62,12 +82,17 @@ class FileUtils {
 
       List<UploadFile> imagesAsByte = [];
       for (var xFile in images) {
-        final imgByte = await xFile.readAsBytes();
+        final isWithinLimit = FileValidationHelper.isWithinLimit(
+          await xFile.length(),
+        );
+        final imgByte = isWithinLimit ? await xFile.readAsBytes() : null;
+
         imagesAsByte.add(
           UploadFile(
-            bytes: await xFile.readAsBytes(),
+            bytes: imgByte,
             name: getFileSignature(xFile.name, imgByte),
-            isMaxSize: FileValidationHelper.isWithinLimit(await xFile.length()),
+            size: await xFile.length(),
+            isMaxSize: !isWithinLimit,
           ),
         );
       }
@@ -103,12 +128,16 @@ class FileUtils {
         return null;
       }
 
-      final imgByte = await image.readAsBytes();
+      final isWithinLimit = FileValidationHelper.isWithinLimit(
+        await image.length(),
+      );
+      final imgByte = isWithinLimit ? await image.readAsBytes() : null;
 
       return UploadFile(
         bytes: imgByte,
         name: getFileSignature(image.name, imgByte),
-        isMaxSize: FileValidationHelper.isWithinLimit(await image.length()),
+        size: await image.length(),
+        isMaxSize: !isWithinLimit,
       );
     } catch (e) {
       debugPrint('Camera picker error: $e');
@@ -134,6 +163,7 @@ class FileUtils {
     try {
       final result = await FilePicker.pickFiles(
         allowMultiple: allowMultiple,
+        type: FileType.any,
         withData: true, // Important for Web
       );
 
@@ -141,11 +171,15 @@ class FileUtils {
 
       return result.files
           .map((pFIle) {
-            final imgByte = pFIle.bytes;
+            final isWithinLimit = FileValidationHelper.isWithinLimit(
+              pFIle.size,
+            );
+            final imgByte = isWithinLimit ? pFIle.bytes : null;
             return UploadFile(
               bytes: imgByte,
               name: getFileSignature(pFIle.name, imgByte),
-              isMaxSize: FileValidationHelper.isWithinLimit(pFIle.size),
+              size: pFIle.size,
+              isMaxSize: !isWithinLimit,
             );
           })
           .take(maxFiles)
@@ -161,8 +195,14 @@ class UploadFile {
   final Uint8List? bytes;
   final String name;
   bool isMaxSize;
+  int size;
 
-  UploadFile({required this.bytes, required this.name, this.isMaxSize = false});
+  UploadFile({
+    required this.bytes,
+    required this.name,
+    required this.size,
+    this.isMaxSize = false,
+  });
 
   bool get isImage {
     final ext = name.toLowerCase();