+++ lib/swift_agents.dart
@@ -1,10 +1,5 @@
 library swift_agents;
 
 export 'src/theme/theme.dart';
-// export 'src/screens/messages_screen.dart';
-// export 'src/widgets/sidebar.dart';
 export 'src/screens/swift_agents_view.dart';
-// export 'src/widgets/chat_bubble.dart';
-// export 'src/widgets/chat_input.dart';
-// export 'src/widgets/top_bar.dart';
 export 'src/swift_agents_sdk.dart';